Add files via upload
The website is designed using HTML and CSS, with the help of the Bootstrap framework for layout and styling. Bootstrap provides a grid system, typography, forms, buttons, and other UI components that make it easy to create responsive and mobile-friendly websites.
Overall, this portfolio project demonstrates the developer's proficiency in front-end web development, including HTML, CSS, and Bootstrap. It also showcases the developer's creativity, design skills, and attention to detail.